repo-marketing/src/requirements.txt
functions-framework==3.*
pandas-gbq==0.23.1
google-cloud==0.34.0